fix: 일정 관련 메소드 각 기능에 맞게 위치 변경
