+++ homecare/src/main/java/jaega/homecare/domain/caregiver/repository/CertificationRepository.java
@@ -10,7 +10,5 @@ import java.util.UUID;
 public interface CertificationRepository extends JpaRepository<Certification, Long> {
     Optional<Certification> findByCertificationId(UUID certificationId);
 
-    boolean existsByCaregiver(Caregiver caregiver);
-
     Optional<Certification> findByCaregiver(Caregiver caregiver);
 }
